+++ lib/widgets/scroll_both_ways.dart
@@ -4,17 +4,21 @@ class ScrollBothWay extends StatelessWidget {
   final ScrollController _horizontal = ScrollController();
   final ScrollController _vertical = ScrollController();
   final Widget child;
+  final double width;
+  final double height;
 
   ScrollBothWay({
     super.key,
     required this.child,
+    this.height = 2400,
+    this.width = 2400,
   });
 
   @override
   Widget build(final BuildContext context) {
     return SizedBox(
-      height: 2500,
-      width: 2400,
+      height: height,
+      width: width,
       child: Scrollbar(
         controller: _vertical,
         thumbVisibility: true,